Need to use Longs instead of Ints

+++ 2023/day05_ifyougiveaseedafertilizer/src/main/scala/AlmanacAnalyzer.scala
@@ -1,5 +1,5 @@
 object AlmanacAnalyzer {
-  def findLowestLocation(almanac: List[String]): Int = {
+  def findLowestLocation(almanac: List[String]): Long = {
     val seeds = SeedParser.parseSeeds(almanac)
     val maps = AlmanacMapParser.parseMaps(almanac)
     val finalValues = AlmanacPipeline.processPipeline(seeds, maps)
